chore: implementação da arquitetura limpa com Angular 18
