+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "restaurantto-app",
-  "version": "0.0.4",
+  "version": "0.0.5",
   "description": "Restaurantto App - Front-end System",
   "type": "module",
   "private": true,
