+++ src/app/app.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { NavbarComponent } from '@lib/components';
+import { NavbarComponent } from '@presentation/view/components';
 
 
 @Component({
