+++ src/app/domain/base/entity.ts
@@ -1,0 +1,10 @@
+export abstract class BaseEntity {
+    constructor(
+        protected id: string,
+        protected createdAt: Date,
+    ) {}
+
+    getId(): string {
+        return this.id;
+    }
+}
