+++ src/app/domain/base/use-case.ts
@@ -1,0 +1,3 @@
+export interface UseCase<Request, Response> {
+    execute(request: Request): Promise<Response>;
+}
