+++ src/app/domain/interfaces/homepage/homepage.ts
@@ -1,0 +1,27 @@
+export interface Homepage {
+    intro: {
+        title: string;
+        description: string;
+        button: {
+            text: string;
+            link: string;
+        };
+        image: string;
+    };
+    cardsContent: {
+        id: number;
+        icon: string;
+        title: string;
+        description: string;
+    }[];
+    section: {
+        title: string;
+        description: string;
+        button: {
+            text: string;
+            link: string;
+        };
+        image: string;
+        reverse: string;
+    };
+}
