+++ src/app/domain/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './homepage/homepage';
