+++ src/app/infra/services/index.ts
@@ -1,0 +1,1 @@
+export * from './data-transfer/data-transfer.service'
