--- src/app/lib/layouts/homepage-layout/section-layout/section-layout.component.ts
@@ -1,19 +1,0 @@
-import { NgClass } from '@angular/common';
-import { Component, Input } from '@angular/core';
-import { ButtonComponent } from '@lib/components';
-
-@Component({
-    selector: 'app-section-layout',
-    standalone: true,
-    templateUrl: './section-layout.component.html',
-    styles: [],
-    imports: [ButtonComponent, NgClass],
-})
-export class SectionLayoutComponent {
-    @Input() title: string = '';
-    @Input() description: string = '';
-    @Input() buttonText: string = '';
-    @Input() buttonLink: string = '';
-    @Input() imageUrl: string = '';
-    @Input() reverse: boolean = false;
-}
