--- src/app/lib/layouts/index.ts
@@ -1,2 +1,0 @@
-export * from './homepage-layout/homepage-layout.component'
-export * from './homepage-layout/banner-layout/banner-layout.component'
