--- src/app/lib/services/index.ts
@@ -1,1 +1,0 @@
-export * from './data-transfer/data-transfer.service';
