--- src/app/pages/homepage/homepage.component.html
@@ -1,3 +1,0 @@
-<app-homepage-layout>
-    
-</app-homepage-layout>
