--- src/app/pages/homepage/homepage.component.ts
@@ -1,13 +1,0 @@
-import { Component } from '@angular/core';
-import { HomepageLayoutComponent } from '@lib/layouts';
-
-
-@Component({
-    selector: 'app-homepage',
-    standalone: true,
-    imports: [HomepageLayoutComponent],
-    templateUrl: './homepage.component.html',
-    styles: ``,
-})
-export class Homepage {
-}
