+++ src/app/presentation/view/components/card/card.component.html
