+++ src/app/presentation/view/layouts/homepage-layout/banner-layout/banner-layout.component.spec.ts
