+++ src/app/presentation/view/layouts/homepage-layout/section-layout/section-layout.component.html
@@ -1,0 +1,12 @@
+<section class="w-full flex flex-wrap xl:flex-nowrap lg:h-screen p-8 lg:p-16 min-h-screen justify-between items-center my-32 lg:my-48">
+  <div [ngClass]="{'xl:order-2 justify-end': reverse, 'xl:order-1 justify-start': !reverse}" class="w-full xl:w-1/2 flex justify-center items-center p-4 my-4 lg:my-8">
+    <img [src]="imageUrl" alt="Imagem ilustrativa da seção" class="w-full lg:w-[50%] h-auto">
+  </div>
+  <div [ngClass]="{'xl:order-1': reverse, 'xl:order-2 justify-end': !reverse}" class="w-full xl:w-1/2 p-4 lg:p-6 my-4 lg:my-8">
+    <h1 class="font-bold text-2xl lg:text-4xl mb-4">{{ title }}</h1>
+    <p class="text-lg lg:text-xl mb-4">{{ description }}</p>
+    <app-button class="w-full lg:w-1/3 py-2 lg:py-4 bg-[rgba(116,3,24,0.73)] text-white rounded-xl" [link]="buttonLink">
+      {{ buttonText }}
+    </app-button>
+  </div>
+</section>
