+++ src/app/presentation/view/layouts/homepage-layout/section-layout/section-layout.component.spec.ts
