+++ src/app/presentation/view/layouts/index.ts
@@ -1,0 +1,3 @@
+export * from './homepage-layout/homepage-layout.component';
+export * from './homepage-layout/banner-layout/banner-layout.component';
+export * from './homepage-layout/section-layout/section-layout.component'
