+++ src/app/presentation/view/pages/homepage/homepage.component.html
@@ -1,0 +1,1 @@
+<app-homepage-layout></app-homepage-layout>
