+++ src/app/presentation/view/pages/homepage/homepage.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { HomepageLayoutComponent } from '@presentation/view/layouts';
+
+
+@Component({
+    selector: 'app-homepage',
+    standalone: true,
+    imports: [HomepageLayoutComponent],
+    templateUrl: './homepage.component.html',
+    styles: ``,
+})
+export class HomepageComponent {
+}
